Rename types directory
