+++ src/__types__/GraphElements.ts
